Complete auth card
