+++ reactproject/src/App.css
@@ -537,6 +537,20 @@ p {
     align-items: center;
 }
 
+.auth-card {
+    margin: 200px auto;
+    width: 69%;
+}
+
+.auth-card .wide-btn {
+    margin-top: 15px;
+}
+
+.card-logo {
+    font-size: 50px;
+    font-weight: bold;
+}
+
 @media screen and (max-width: 1200px) {
     .wide-btn {
         width: 100px;
