+++ reactproject/src/pages/Main.jsx
@@ -1,15 +1,19 @@
 import React from "react";
 import CalendarSection from "../components/calendarSection/CalendarSection.jsx";
 import UserSection from "../components/userSection/UserSection.jsx";
+import Header from '../components/header/Header'
 
 const Main = () => {
     return (
-        <main className="main">
-            <div className="container">
-                <UserSection />
-                <CalendarSection />
-            </div>
-        </main>
+        <>
+            <Header />
+            <main className="main">
+                <div className="container">
+                    <UserSection />
+                    <CalendarSection />
+                </div>
+            </main>
+        </>
     )
 }
